Allow unneccesary_cast for the Specifier implementation.
When a bitfield struct is annotated with `derive(BitfieldSpecifier)`, the
`Specifier` trait is implemented for the type.

In the implementation of the trait, under `from_bytes`, a literal is cast to
`Self::Bytes`, which contains the type of the required integer representation.

The casting is identified as unnecessary by clippy, when expanded, but seems to
be necessary to avoid an ambiguous call.

For this reasons, `unnecessary_cast` was allowed in the code block where this happened.